@@ -31,6 +31,7 @@ import 'package:pdf/pdf.dart';
 import 'callback.dart';
 import 'interface.dart';
 import 'method_channel_ffi.dart' if (dart.library.js) 'method_channel_js.dart';
+import 'output_type.dart';
 import 'print_job.dart';
 import 'printer.dart';
 import 'printing_info.dart';
@@ -77,14 +78,16 @@ class MethodChannelPrinting extends PrintingPlatform {
             return true;
           }());
 
-          FlutterError.reportError(FlutterErrorDetails(
-            exception: e,
-            stack: s,
-            stackFilter: (input) => input,
-            library: 'printing',
-            context: ErrorDescription('while generating a PDF'),
-            informationCollector: collector,
-          ));
+          FlutterError.reportError(
+            FlutterErrorDetails(
+              exception: e,
+              stack: s,
+              stackFilter: (input) => input,
+              library: 'printing',
+              context: ErrorDescription('while generating a PDF'),
+              informationCollector: collector,
+            ),
+          );
 
           if (job.useFFI) {
             return setErrorFfi(job, e.toString());
@@ -178,6 +181,7 @@ class MethodChannelPrinting extends PrintingPlatform {
     PdfPageFormat format,
     bool dynamicLayout,
     bool usePrinterSettings,
+    OutputType outputType,
   ) async {
     final job = _printJobs.add(
       onCompleted: Completer<bool>(),
@@ -196,6 +200,7 @@ class MethodChannelPrinting extends PrintingPlatform {
       'marginBottom': format.marginBottom,
       'dynamic': dynamicLayout,
       'usePrinterSettings': usePrinterSettings,
+      'outputType': outputType.index,
     };
 
     await _channel.invokeMethod<int>('printPdf', params);
@@ -230,7 +235,9 @@ class MethodChannelPrinting extends PrintingPlatform {
       'h': bounds.height,
     };
     final printer = await _channel.invokeMethod<Map<dynamic, dynamic>>(
-        'pickPrinter', params);
+      'pickPrinter',
+      params,
+    );
     if (printer == null) {
       return null;
     }
@@ -262,7 +269,10 @@ class MethodChannelPrinting extends PrintingPlatform {
 
   @override
   Future<Uint8List> convertHtml(
-      String html, String? baseUrl, PdfPageFormat format) async {
+    String html,
+    String? baseUrl,
+    PdfPageFormat format,
+  ) async {
     final job = _printJobs.add(
       onHtmlRendered: Completer<Uint8List>(),
     );
